requirements.txt
bcrypt==3.1.7
cffi==1.12.3
Django==2.2.5
djangorestframework==3.10.3
pycparser==2.19
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
